+++ tests/testable/mailer/swift.php
@@ -1,0 +1,9 @@
+<?php namespace Orchestra\Testable\Mailer;
+
+class Swift {
+
+	public function getFrom()
+	{
+		return ;
+	}
+}